feat: User update, delete API 예외 처리
사용자의 존재 여부를 확인 후 update, delete 수행하도록 변경

+++ src/main/java/com/group/libraryapp/controller/user/UserController.java
@@ -18,7 +18,7 @@ public class UserController {
 
     private final JdbcTemplate jdbcTemplate;
 
-    public UserController(JdbcTemplate jdbcTemplate){
+    public UserController(JdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
     }
 
@@ -50,13 +50,35 @@ public class UserController {
 
 
     @PutMapping("/user")
-    public void updateUser(@RequestBody UserUpdateRequest request){
+    public void updateUser(@RequestBody UserUpdateRequest request) {
+
+        // User 존재 여부 확인
+        String readSql = "SELECT * FROM user WHERE id = ?";
+        boolean isUserNotExist = jdbcTemplate
+                .query(readSql, (rs, rowNum) -> 0, request.getId())
+                .isEmpty();
+
+        if (isUserNotExist) {
+            throw new IllegalArgumentException();
+        }
+
         String sql = "UPDATE user SET name = ? WHERE id = ?";
         jdbcTemplate.update(sql, request.getName(), request.getId());
     }
 
     @DeleteMapping("/user")
-    public void deleteUser(@RequestParam String name){
+    public void deleteUser(@RequestParam String name) {
+
+        // User 존재 여부 확인
+        String readSql = "SELECT * FROM user WHERE name = ?";
+        boolean isUserNotExist = jdbcTemplate
+                .query(readSql, (rs, rowNum) -> 0, name)
+                .isEmpty();
+
+        if (isUserNotExist) {
+            throw new IllegalArgumentException();
+        }
+
         String sql = "DELETE FROM user WHERE name = ?";
         jdbcTemplate.update(sql, name);
     }
